Remove unnecessary git output
