+++ fixtures/example_dot/Dot.toml
@@ -1,5 +1,5 @@
 [package]
-name = "example_package"
+name = "example_dot"
 authors = [ "Michael Mullins" ]
 
 [link]
