+++ test_utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+camino = "1.0.5"
+failure = "0.1.8"
+pretty_assertions = "1.0.0"
+tempfile = "3.2.0"
+walkdir = "2.3.2"
