+++ tests/output/add_dot_success.out
@@ -1,6 +1,3 @@
 [info] Adding {SRC_PATH}
 [info] Cloning...
-Cloning into '{DEST_PATH}'...
-warning: --depth is ignored in local clones; use file:// instead.
-done.
 [info] Cloning complete!
